Moved Castle Drakan steps into its own class
It added a lot of noise to the main helper, so hopefully this'll help.
